More WIP progress to the basevar generic refactor. NumberVariables are now doubles internally and VectorVariables are now Vector4 types.

+++ Runtime/Operations/NumberOperation.cs
@@ -75,14 +75,14 @@ namespace Buck
                 
                 case TypeCode.Single:
                     FloatVariable floatA = (FloatVariable)m_numberA;
-                    floatA.Value = (decimal)GetFloatResult();
+                    floatA.Value = GetFloatResult();
 
                     break;
 
                 
                 case TypeCode.Double:
                     DoubleVariable doubleA = (DoubleVariable)m_numberA;
-                    doubleA.Value = (decimal)GetDoubleResult();
+                    doubleA.Value = GetDoubleResult();
                     break;
             }
